# contracts/final-cut-review/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Final Cut Review Contract
  version: "1.0"
  x-contract-version: "1.0"
servers:
  - url: /api/v1
...
security:
  - PrincipalContext: []
  - WriteGuardCookie: []
  - ReverseProxyWriteGuard: []
paths:
...
  /final-cut-review/write-guard/session:
    post:
      operationId: verify_write_guard
      tags: [commands]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedCodeRequest"
      responses:
        "200":
          description: Write guard verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeWriteGuardSession"
        <<: *standardNon200Responses
...
components:
  securitySchemes:
    PrincipalContext:
      type: apiKey
      in: header
      name: X-Principal-Context
      description: Signed principal context token issued by the trusted host.
    WriteGuardCookie:
      type: apiKey
      in: cookie
      name: fj_write_guard
    ReverseProxyWriteGuard:
      type: apiKey
      in: header
      name: X-Write-Guard-Verified
...
  schemas:
...
    ResponseMeta:
      type: object
      required: [request_id, contract_version]
      additionalProperties: false
      properties:
        request_id: {type: string}
        contract_version: {type: string, const: "1.0"}
        total_count: {type: integer, minimum: 0}
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1}
...
    EnvelopeAny:
      type: object
      required: [data, meta]
      additionalProperties: false
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    EnvelopeWriteGuardSession:
      allOf:
        - $ref: "#/components/schemas/EnvelopeAny"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/WriteGuardSessionDTO"
...
    SharedCodeRequest:
      type: object
      required: [code]
      additionalProperties: false
      properties:
        code: {type: string, minLength: 1, maxLength: 256}
    WriteGuardSessionDTO:
      type: object
      required: [verified, mode]
      additionalProperties: false
      properties:
        verified: {type: boolean}
        mode: {type: string, enum: [shared_code, reverse_proxy, none]}
...