# src/modules/final-cut-review/contracts-generated/backend-contract-models.ts
// Generated by backend/scripts/generate_contracts.py. Do not edit by hand.
import type { CommandType, ErrorCode, ErrorHttpStatus } from "./backend-contract-metadata";

export interface ErrorBody {
  code: ErrorCode;
  message: string;
  http_status: ErrorHttpStatus;
  details: Record<string, unknown>;
  request_id: string;
  timestamp: string;
  contract_version: "1.0";
}

export interface ResponseMeta {
  request_id: string;
  contract_version: "1.0";
  total_count?: number;
  page?: number;
  page_size?: number;
}

export interface ListResponseMeta {
  request_id: string;
  contract_version: "1.0";
  total_count: number;
  page: number;
  page_size: number;
}

export interface Envelope<T> {
  data: T;
  meta: ResponseMeta;
}

export interface ErrorEnvelope {
  error: ErrorBody;
}

export interface CommandEnvelope<TPayload extends object> {
  command_id: string;
  command_type: CommandType;
  contract_version: "1.0";
  expected_aggregate_version?: number;
  payload: TPayload;
}

export interface ReviewAnnotationShape {
  id: string;
  tool_type: "pen" | "arrow" | "rect" | "circle" | "text";
  anchor_points?: Array<{ x: number; y: number }>;
  path_data?: string;
  text_content?: string;
  color: string;
  line_width: number;
  font_size?: number;
  z_index: number;
}

export interface ReviewAnnotationSetInput {
  canvas_width: number;
  canvas_height: number;
  video_width: number;
  video_height: number;
  shapes: ReviewAnnotationShape[];
}

export interface ReviewPlaybackTarget {
  project_ref_id: string;
  review_item_id: string;
  version_id: string;
  issue_id: string;
  revision_id: string;
  annotation_set_id?: string;
  timestamp_ms: number;
  frame_number: number;
}

export interface OriginalMediaSnapshotDTO {
  original_file_id: string;
  original_filename: string;
  mime_type: string;
  file_size: number;
  sha256: string;
  duration_ms: number;
  width: number;
  height: number;
  fps_num: number;
  fps_den: number;
  media_probe_version: string;
}

export interface ProjectDTO {
  project_ref_id: string;
  project_code: string;
  project_name: string;
  description: string;
  source: "local" | "host";
  external_project_id?: string | null;
  lifecycle_status: "active" | "archived";
  completion_status: "empty" | "in_progress" | "completed";
  deleted_at?: string | null;
  lock_version: number;
  created_at: string;
  updated_at: string;
}

export interface ReviewVersionDTO {
  id: string;
  project_ref_id: string;
  review_item_id: string;
  previous_version_id?: string | null;
  version_no: number;
  version_label: string;
  is_current: boolean;
  original_media: OriginalMediaSnapshotDTO;
  playback_status: "processing" | "ready" | "failed";
  playback_asset_id?: string | null;
  thumbnail_asset_id?: string | null;
  version_note?: string | null;
  change_summary?: string | null;
  lock_version: number;
  created_at: string;
}

export interface ReviewIssueRevisionDTO {
  id: string;
  project_ref_id: string;
  review_item_id: string;
  version_id: string;
  issue_id: string;
  revision_no: number;
  content: string;
  annotation_set_id?: string | null;
  created_at: string;
}

export interface ReviewAnnotationSetDTO {
  id: string;
  project_ref_id: string;
  review_item_id: string;
  version_id: string;
  issue_id: string;
  timestamp_ms: number;
  frame_number: number;
  canvas_width: number;
  canvas_height: number;
  video_width: number;
  video_height: number;
  shapes: ReviewAnnotationShape[];
  created_at: string;
}

export interface ReviewIssueDTO {
  id: string;
  project_ref_id: string;
  review_item_id: string;
  version_id: string;
  issue_no: number;
  status: "unresolved" | "resolved";
  current_revision_id: string;
  timestamp_ms: number;
  frame_number: number;
  playback_target: ReviewPlaybackTarget;
  current_revision: ReviewIssueRevisionDTO;
  current_annotation_set?: ReviewAnnotationSetDTO | null;
  deleted_at?: string | null;
  lock_version: number;
  created_at: string;
  updated_at: string;
}

export interface ReviewItemDTO {
  id: string;
  project_ref_id: string;
  item_code: string;
  episode_no?: number | null;
  title: string;
  workflow_status: "pending_review" | "in_review" | "changes_requested" | "finalized";
  current_version_id: string;
  current_version_no: number;
  ui_status: string;
  active_finalization_id?: string | null;
  unresolved_current_version_count: number;
  resolved_current_version_count: number;
  historical_version_count: number;
  is_finalized: boolean;
  lock_version: number;
  created_at: string;
  updated_at: string;
}

export interface ThreadMessageDTO {
  id: string;
  project_ref_id: string;
  review_item_id: string;
  version_id: string;
  issue_id: string;
  content: string;
  created_at: string;
}

export interface FinalizationDTO {
  id: string;
  project_ref_id: string;
  review_item_id: string;
  version_id: string;
  version_no: number;
  original_media: OriginalMediaSnapshotDTO;
  status: "active";
  finalized_at: string;
}

export interface PackageSnapshotItem {
  review_item_id: string;
  version_id: string;
  finalization_id: string;
  original_file_id: string;
  original_filename: string;
  sha256: string;
  archive_name: string;
}

export interface PackageSnapshotDTO {
  id: string;
  project_ref_id: string;
  status: "preparing" | "ready" | "failed" | "expired";
  package_filename: string;
  expires_at: string;
  file_count: number;
  total_bytes: number;
  sha256?: string;
  download_token?: string;
  download_token_expires_at?: string;
  created_at: string;
  updated_at: string;
  failure_details?: Record<string, unknown>;
  items: PackageSnapshotItem[];
}

export interface PackageDownloadSessionDTO {
  status: "ready";
}

export interface UploadInitRequest {
  original_filename: string;
  mime_type: "video/mp4" | "video/quicktime" | "application/octet-stream";
  file_size: number;
  sha256: string;
  duration_ms?: number;
  width?: number;
  height?: number;
  fps_num?: number;
  fps_den?: number;
}

export interface UploadSessionDTO {
  upload_id: string;
  status: "initiated" | "receiving" | "completed" | "aborted";
  original_filename: string;
  mime_type: string;
  declared_size: number;
  received_size: number;
  file_id?: string | null;
}